PR feedback: Move tests, make test type generic
