+++ src/libraries/System.Runtime.InteropServices/tests/ComInterfaceGenerator.Unit.Tests/VerifyCompilationTest.cs
@@ -2,19 +2,15 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
 using Microsoft.Interop.UnitTests;
-using VerifyCS = Microsoft.Interop.UnitTests.Verifiers.CSharpSourceGeneratorVerifier<Microsoft.Interop.ComClassGenerator>;
 
 namespace ComInterfaceGenerator.Unit.Tests
 {
-    internal class VerifyCompilationTest : VerifyCS.Test
+    internal class VerifyCompilationTest<T> : Microsoft.Interop.UnitTests.Verifiers.CSharpSourceGeneratorVerifier<T>.Test
+        where T : new()
     {
-        public Action<Compilation>? CompilationVerifier { get; init; }
+        public required Action<Compilation> CompilationVerifier { get; init; }
 
         public VerifyCompilationTest(TestTargetFramework targetFramework) : base(targetFramework)
         {
@@ -24,10 +20,6 @@ namespace ComInterfaceGenerator.Unit.Tests
         {
         }
 
-        protected override void VerifyFinalCompilation(Compilation compilation)
-        {
-            if(CompilationVerifier != null)
-                CompilationVerifier(compilation);
-        }
+        protected override void VerifyFinalCompilation(Compilation compilation) => CompilationVerifier(compilation);
     }
 }
